Added notes
Added notes that the player will be able to pick up. As properly picking them up will require scene changes to add them, changes to the player script to register them, and changed to the pause menu to view them, their addition as a system will be done when we do not need to do a presentation of our progress.