Added Return Point to Blackguard
Gave the Blackguard a return point, which it currently instantly travels to, so that it does not just stop mid air when the player leave's it's zone. Ideally, it should instead travel to that point and then stop trying to reach it when it gets close enough to reduce lag. Also, began working on the player's Lighter mode for after they lose their phone.